+++ src/main/kotlin/org/jetbrains/intellij/IntelliJPlugin.kt
@@ -1,6 +1,7 @@
 package org.jetbrains.intellij
 
 import com.jetbrains.plugin.structure.intellij.version.IdeVersion
+import org.gradle.api.GradleException
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.api.artifacts.Configuration
@@ -1263,7 +1264,7 @@ open class IntelliJPlugin : Plugin<Project> {
 
                         else -> {
                             if (extension.version.orNull != null) {
-                                warn(context, "Both 'localPath' and 'version' specified, second would be ignored")
+                                throw GradleException("Both 'localPath' and 'version' specified, second would be ignored")
                             }
                             info(context, "Using path to locally installed IDE: $localPath")
                             dependencyManager.resolveLocal(project, localPath, extension.localSourcesPath.orNull)