+++ CHANGES.md
@@ -9,6 +9,7 @@
 - Integration Tests: Cover `instrumentCode` task
 - Add `java.system.class.loader=com.intellij.util.lang.PathClassLoader` system property for tests run on 221+
 - Fixed version parsing for `listProductsReleases` task which resulted in incorrect IDE releases versions
+- Throw an exception instead of warning when both `localPath` and `version` are specified
 
 ## 1.5.3
 - Updated dependencies marked as vulnerable [#459](../../issues/459)